translators-master/requirements.txt
azure-ai-textanalytics==5.3.0
azure-ai-translation-text==1.0.0b1
azure-common==1.1.28
azure-core==1.29.3
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dill==0.3.7
et-xmlfile==1.1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isodate==0.6.1
libretranslatepy==2.1.1
lxml==4.9.3
msrest==0.7.1
multiprocess==0.70.15
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.3
pathos==0.3.1
pox==0.3.3
ppft==1.7.6.7
pycparser==2.21
PyExecJS==1.5.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
translate==3.6.1
translators==5.8.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
XlsxWriter==3.1.2
